Now Mozart compile mostly warning-free again! + Fixed most warnings about missing const with literal strings (remaining one is for virtualString to CString which is to be reviewed) + Changed the reporting flags in the DSS to sane values (removed -Weffc++ and -Winline, forced the use of most flags for GCC > 3) + Fixed most of the remaining warnings but   - usage of tmpnam -> requires some user-visible changes   - warnings in ozbison; well it's in code that's more than 12(!) years old. The good way would be to update the underlying bison. Pain ahead. + Fixed some bugs here and there (warnings are supposed to help locate them when they aren't drown in noise) +The output of make bootstrap dropped from 1.7 Mb to 189 kb!

+++ platform/dss/src/dss_msgLayerInterface.hh
@@ -183,7 +183,7 @@ namespace _dss_internal{ //Start namespace
     SimpleBlockBuffer *a_curBuf; 
     BYTE *a_cur; 
   public:
-    virtual int  availableSpace() const;
+    virtual size_t  availableSpace() const;
     virtual bool canWrite(size_t len) const {return this->availableSpace()>=len;}
     virtual void writeToBuffer(const BYTE* ptr, size_t len);
     virtual void putByte(const BYTE& b);
